requirements.txt
langchain==0.1.11
langchain-openai==0.0.8
openai==1.13.3
pydantic==2.5.3
python-dotenv==1.0.0